feat: add PDF/X validator family (X-1a, X-3, X-4, X-4p, X-6, X-6p)
Adds nine new Validator variants covering the PDF/X prepress standards:
X-1a (ISO 15930-4), X-3 (ISO 15930-6), X-4/X-4p (ISO 15930-7), X-6/X-6p
(ISO 15930-9), plus combined PDF/A + PDF/X validators (A1B_X1A, A2B_X4,
A3B_X4). X-4p and X-6p support external ICC profile references via the
new ExternalOutputProfile type, whose rgb/luma/cmyk constructors return
Result<_, ExternalOutputProfileError> and validate required fields
(non-empty URLs, identifier, info) eagerly at construction time.

Four new ValidationError variants (ContainsRgb, MixedGradientColorSpaces,
MissingTrimOrArtBox, ContainsAnnotation) plus ExternalOutputProfileRequiresX4P
enforce PDF/X-specific requirements. A new public Trapping enum
(Trapped/NotTrapped/Unknown) replaces the previous bool for Metadata::trapped;
PDF/X forbids Unknown and krilla downgrades it to NotTrapped when a PDF/X
validator is active. Output intent serialization was refactored to support
multiple intents for combined validators, with CMYK output intents shared
across the PDF/A and PDF/X subtypes as permitted by ISO 15930-7 and
ISO 19005-2. The /Trapped key is now always written to the Document Info
dictionary for PDF/X, including PDF 2.0-based X-6/X-6p (ISO 15930-9 still
mandates it regardless of the general PDF 2.0 Info-dict deprecation), and
the XMP path is in sync via Metadata::serialize_xmp_metadata. /GTS_PDFXVersion
is emitted for every PDF/X revision in both the Info dict and XMP metadata.

The ICC profile parser now uses checked_add in the three remaining
arithmetic paths that were previously unguarded (parse_profile_name,
parse_desc_profile_description, parse_mluc_profile_description), making
ICCProfile::new safe to call on arbitrary untrusted bytes. Gradient
serialization now handles empty stops lists gracefully via an early
skip rather than panicking on .expect().

+++ crates/krilla/examples/parley.rs
@@ -129,7 +129,7 @@ fn main() {
 
                     // Add the current glyph to our buffer of glyphs.
                     glyphs.push(KrillaGlyph::new(
-                        GlyphId::new(glyph.id as u32),
+                        GlyphId::new(glyph.id),
                         glyph.advance / font_size,
                         glyph.x / font_size,
                         glyph.y / font_size,
